Reorganized tokenizer in multiple classes
